Explicit check and documentation for negative values
We could try and be clever and allow the FWHM to be estimated
if the data is negative, but it's tricky to know what to do
when the values switch between positive and negative. So
we are instead explicit that this is only intended for positive
values and if negative ones are sent in then the fall-through
case (half the x range) is used.